dont delete main branches

+++ src/github/logic.py
@@ -301,7 +301,8 @@ def maybe_delete_branch_if_merged(pull_request: PullRequest):
         owner = pull_request.repository_owner_handle()
         repo_name = pull_request.repository_name()
         branch_name = pull_request.head_ref_name()
-        github_client.delete_branch_if_exists(owner, repo_name, branch_name)
+        if branch_name not in ["master", "main"]:
+            github_client.delete_branch_if_exists(owner, repo_name, branch_name)
 
 
 # ----------------------------------------------------------------------------------
